requirements.txt
pandas>=1.5,<3.0
numpy>=1.22,<2.0
scikit-learn>=1.1,<1.5
torch>=2.0,<3.0
transformers>=4.31,<5.0
PyYAML>=6.0,<7.0
vectorbt>=0.25,<0.26
nltk>=3.8,<4.0
